+++ src/main/java/com/inet/sass/function/LightenFunctionGenerator.java
@@ -39,8 +39,8 @@ class LightenFunctionGenerator extends AbstractFunctionGenerator {
             LexicalUnitImpl function, FormalArgumentList actualArguments) {
         LexicalUnitImpl color = getParam(actualArguments, "color")
                 .getContainedValue();
-        float amount = getParam(actualArguments, "amount").getContainedValue()
-                .getFloatValue();
+        double amount = getParam(actualArguments, "amount").getContainedValue()
+                .getDoubleValue();
         return ColorUtil.lighten(color, amount);
     }
